Eliminate unnecessary utility class
